Programa corrigido

+++ main/java/menuExtensao/VeiculoMenu.java
@@ -3,6 +3,7 @@ package menuExtensao;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+import br.com.mildevs.Program;
 import br.com.mildevs.dao.CondutorDAO;
 import br.com.mildevs.dao.MultaDAO;
 import br.com.mildevs.dao.VeiculoDAO;
@@ -12,7 +13,7 @@ import br.com.mildevs.entity.Veiculo;
 
 public class VeiculoMenu {
 
-	private static final Condutor Condutor = null;
+	private static final String placa = null;
 	public static Scanner tc = new Scanner(System.in);
 	private static VeiculoDAO veiculoDAO = new VeiculoDAO();
 	private static CondutorDAO condutorDAO = new CondutorDAO();
@@ -50,10 +51,11 @@ public class VeiculoMenu {
 				break;
 
 			case 5:
-				// venderVeiculo();
+				venderVeiculo();
 				break;
 
 			case 6:
+				Program.menuPrincipal();
 				return;
 
 			default:
@@ -63,19 +65,21 @@ public class VeiculoMenu {
 		}
 	}
 
-
-	public static void cadastrarcoVeiculo(Scanner tc)
-	{
+	public static void cadastrarcoVeiculo(Scanner tc) {
 		System.out.println("====== Insira os dados nescessários para Cadastrar o veículo");
 		System.out.println();
 		System.out.print("CNH do condutor: ");
 		int cnh = tc.nextInt();
 		Condutor condutor = CondutorDAO.consultarCondutor((cnh));
 		if (condutor == null) {
-			System.err.println("Este condutor não está cadastrado");
+			System.err.println("Este condutor não está cadastrado"
+					+ "Cadastre o condutor e tente novamente");
+			menu();
 		} else if (condutor.getVeiculo() != null) {
 			System.out.println();
-			System.err.println("Este condutor já possui um veículo cadastrado");
+			System.err.println("Este condutor já possui um veículo cadastrado"+
+			"Revise os dados e tente novamente.");
+			menu();
 		} else {
 
 			Veiculo veiculo = new Veiculo();
@@ -94,22 +98,31 @@ public class VeiculoMenu {
 			veiculo.setMarca(marca);
 			veiculo.setAno(ano);
 			veiculo.setMultas(new ArrayList<Multa>());
-        VeiculoDAO veiculodao = new VeiculoDAO();
-		veiculodao.criarVeiculo(veiculo, condutor);
-			
+			VeiculoDAO veiculodao = new VeiculoDAO();
+			veiculodao.criarVeiculo(veiculo, condutor);
+			menu();
+
 		}
-	
+
 	}
 
 	private static void removerVeiculo() {
 		System.out.println();
-		System.out.println("=== Adicione os dados necessários para remoção ===");
-		System.out.println();
-
+		System.out.println("==== Insira os dados nescessários para a remoção do veiculo:");
 		System.out.print("Placa: ");
 		String placa = tc.next();
+		Veiculo veiculo = VeiculoDAO.consultarVeiculos(placa);
+		if (veiculo == null) {
+			System.out.println("Veiculo nao encontrado, confira os dados e tente novamente");
+			menu();
+		} else if (veiculo.getMulta() != null) {
+			System.out.println("Não é possivel remover veiculos com multas.");
+			menu();
 
-		veiculoDAO.removeVeiculo(placa);
+		} else {
+			veiculoDAO.removeVeiculo(placa);
+			menu();
+		}
 	}
 
 	private static void consultarVeiculo() {
@@ -117,14 +130,16 @@ public class VeiculoMenu {
 		System.out.println("==== Adicione os dados necessários para a consulta ====");
 		System.out.print("Placa: ");
 		String placa = tc.next();
+		Veiculo veiculo = VeiculoDAO.consultarVeiculos(placa);
 
-		Veiculo veiculo = veiculoDAO.consultarVeiculos(placa);
 		if (veiculo == null) {
 			System.out.println();
-			System.err.println("Nenhum veículo encontrado");
-
+			System.err.println("Nenhum veículo encontrado, confira os dados e tente novamente.");
+			menu();
 		} else {
 			System.out.println(veiculo);
+			System.out.println("Você será retornado ao menu veiculos");
+			menu();
 		}
 	}
 
@@ -136,15 +151,67 @@ public class VeiculoMenu {
 
 		if (veiculos == null || veiculos.isEmpty()) {
 			System.out.println();
-			System.err.println("Nenhum veículo encontrado");
-
+			System.err.println("Nenhum veículo encontrado, confira os dados e tente novamente");
+			menu();
 		} else {
 			for (Veiculo v : veiculos) {
 				System.out.println(v);
+				menu();
 			}
 
 		}
 
+	}
+
+	private static void venderVeiculo() {
+		System.out.println();
+		System.out.println("Vender veículo");
+		System.out.println();
+
+		System.out.print("CNH do vendedor: ");
+		int origem = tc.nextInt();
+
+		System.out.print("CNH do comprador: ");
+		int destino = tc.nextInt();
+
+		Condutor vendedor = condutorDAO.consultarCondutor(origem);
+		Condutor comprador = condutorDAO.consultarCondutor(destino);
+
+		if (vendedor == null || comprador == null) {
+			System.out.println();
+			System.err.println("Condutor não encontrado, confira os dados e tente novamente.");
+			menu();
+		} else if (vendedor.equals(comprador)) {
+			System.out.println();
+			System.err.println("Vendedor e comprador não podem ser a mesma pessoa");
+			menu();
+
+		} else {
+			System.out.print("Placa do carro: ");
+			String placa = tc.next();
+			Veiculo veiculo = veiculoDAO.consultarVeiculos(placa);
+
+			if (veiculo == null) {
+				System.out.println();
+				System.err.println("Nenhum veículo encontrado");
+				menu();
+			} else {
+				int cnhDono = veiculo.getCondutor().getNroCnh();
+
+				if (cnhDono == (vendedor.getNroCnh())) {
+					System.out.println();
+					System.err.println("Este veículo não pertence ao vendedor");
+					menu();
+				} else if (!(veiculo.getMultas().isEmpty())) {
+					System.out.println();
+					System.err.println("Veículo não pode ser vendido pois tem multas não resolvidas");
+					menu();
+				} else {
+					veiculoDAO.venderVeiculo(vendedor, comprador, veiculo);
+
+				}
+			}
+		}
 	}
 
 }
